# swagger/docs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RAS API",
    "version": "1.0",
    "contact": {
      "name": "RAS gRPC-Gateway project",
      "url": "https://github.com/v8platform/ras-grpc-gw",
      "email": "khorevaa@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/v8paltform/ras-grpc-gw/blob/master/LICENSE"
    }
  },
...
  "host": "localhost:3001",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/login": {
      "post": {
        "summary": "Выполняет вход в систему и возвращает токены доступа",
        "description": "Вход в систему для пользователя",
        "operationId": "UsersService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/service.api.v1.Tokens"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/service.api.v1.LoginRequest"
            }
          }
        ],
        "tags": [
          "Access"
        ],
        "security": []
      }
    },
...
    "/v1/refresh": {
      "post": {
        "summary": "Выполняет обновление токена доступа",
        "description": "Обновление токена доступа для пользователя",
        "operationId": "UsersService_Refresh",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/service.api.v1.Tokens"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/service.api.v1.RefreshRequest"
            }
          }
        ],
        "tags": [
          "Access"
        ],
        "security": [
          {
            "UserAuth": []
          }
        ]
      }
    },
    "/v1/users": {
      "post": {
        "summary": "Выполняет регистрацию нового пользователя в системе",
        "description": "Создает и регистрирует нового пользователя",
        "operationId": "UsersService_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/service.api.v1.CreateUserRequest"
            }
          }
        ],
        "tags": [
          "Users"
        ],
        "security": [
          {
            "AdminAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
...
    "service.api.v1.CreateUserRequest": {
      "type": "object",
      "properties": {
        "user": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "service.api.v1.LoginRequest": {
      "type": "object",
      "example": {
        "user": "admin",
        "password": "123"
      },
      "properties": {
        "user": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "title": "LoginRequest",
      "required": [
        "user",
        "password"
      ]
    },
    "service.api.v1.RefreshRequest": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string"
        }
      }
    },
...
    "service.api.v1.Tokens": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "AdminAuth": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header",
      "x-default": "12312312"
    },
    "ClientAuth": {
      "type": "apiKey",
      "name": "x-client",
      "in": "header",
      "x-default": "12312312"
    },
    "UserAuth": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header",
      "x-default": "12312312"
    }
  },
  "security": [
    {
      "ClientAuth": [],
      "UserAuth": []
    }
  ],
...
}